+++ Crucible/Code/QSFancyTableView.m
@@ -46,7 +46,7 @@
 - (void)drawBackgroundInClipRect:(NSRect)clipRect{
 	
 	NSColor *backgroundColor = [self backgroundColor];
-	float hue, saturation, brightness, alpha;
+	CGFloat hue, saturation, brightness, alpha;
 	[[backgroundColor
       colorUsingColorSpaceName:NSDeviceRGBColorSpace] getHue:&hue
 												  saturation:&saturation brightness:&brightness alpha:&alpha];
@@ -80,7 +80,7 @@
 								//	one row's selection can change how others draw.
 }
 
-- (void)deselectRow:(int)row;
+- (void)deselectRow:(NSInteger)row;
 {
 	[super deselectRow:row];
 	[self setNeedsDisplay:YES]; // we display extra because we draw
@@ -141,7 +141,7 @@
 - (void)drawBackgroundInClipRect:(NSRect)clipRect{
 	
 	NSColor *backgroundColor = [self backgroundColor];
-	float hue, saturation, brightness, alpha;
+	CGFloat hue, saturation, brightness, alpha;
 	[[backgroundColor
       colorUsingColorSpaceName:NSDeviceRGBColorSpace] getHue:&hue
 												  saturation:&saturation brightness:&brightness alpha:&alpha];
@@ -174,7 +174,7 @@
 								//	one row's selection can change how others draw.
 }
 
-- (void)deselectRow:(int)row;
+- (void)deselectRow:(NSInteger)row;
 {
 	[super deselectRow:row];
 	[self setNeedsDisplay:YES]; // we display extra because we draw
