+++ Crucible/Code/QSResultWindow.m
@@ -4,7 +4,8 @@
 
 
 @implementation QSResultWindow
-- (id)initWithContentRect:(NSRect)contentRect styleMask:(unsigned int)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag{
+- (id)initWithContentRect:(NSRect)contentRect styleMask:(NSUInteger)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag
+{
     NSWindow* result = [super initWithContentRect:contentRect styleMask:aStyle backing:bufferingType defer:YES];
     [self setOpaque:![[NSUserDefaults standardUserDefaults]boolForKey:@"QSResultsUseAlpha"]];
