+++ Quicksilver/Plugins/CubeInterface/QSSearchController.m
@@ -1515,7 +1515,7 @@
 }
 //
 - (void)selectObject:(QSBasicObject *)obj {	
-	int index = 0;
+	NSInteger index = 0;
 	//[self updateHistory];
 	if (obj) {
 		index = [resultArray indexOfObject:obj];
